Revert "JAVA-17818 Split or move spring-cloud-openfeign module (conti-2) (#13621)"
This reverts commit 492ef53b72ae83ec29aada73cbb7c06e387752da.
