+++ feign/README.md
@@ -7,3 +7,9 @@ This module contains articles about Feign
 - [Intro to Feign](https://www.baeldung.com/intro-to-feign)
 - [Retrying Feign Calls](https://www.baeldung.com/feign-retry)
 - [Setting Request Headers Using Feign](https://www.baeldung.com/java-feign-request-headers)
+- [File Upload With Open Feign](https://www.baeldung.com/java-feign-file-upload)
+- [Feign Logging Configuration](https://www.baeldung.com/java-feign-logging)
+- [Retrieve Original Message From Feign ErrorDecoder](https://www.baeldung.com/feign-retrieve-original-message)
+- [RequestLine with Feign Client](https://www.baeldung.com/feign-requestline)
+- [Propagating Exceptions With OpenFeign and Spring](https://www.baeldung.com/spring-openfeign-propagate-exception)
+- [Post form-url-encoded Data with Spring Cloud Feign](https://www.baeldung.com/spring-cloud-post-form-url-encoded-data)
