+++ feign/src/main/java/com/baeldung/core/ExampleApplication.java
@@ -2,8 +2,10 @@ package com.baeldung.core;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
+@EnableFeignClients
 public class ExampleApplication {
 
     public static void main(String[] args) {
