+++ feign/src/main/java/com/baeldung/core/customizederrorhandling/config/FeignConfig.java
@@ -1,0 +1,19 @@
+package com.baeldung.core.customizederrorhandling.config;
+
+import org.springframework.context.annotation.Bean;
+
+import feign.Logger;
+import feign.codec.ErrorDecoder;
+
+public class FeignConfig {
+
+    @Bean
+    Logger.Level feignLoggerLevel() {
+        return Logger.Level.FULL;
+    }
+
+    @Bean
+    ErrorDecoder errorDecoder() {
+        return new CustomErrorDecoder();
+    }
+}
