+++ feign/src/main/java/com/baeldung/core/customizederrorhandling/exception/ProductNotFoundException.java
@@ -1,0 +1,8 @@
+package com.baeldung.core.customizederrorhandling.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+}
