+++ feign/src/main/java/com/baeldung/core/customizederrorhandling/exception/ProductServiceNotAvailableException.java
@@ -1,0 +1,8 @@
+package com.baeldung.core.customizederrorhandling.exception;
+
+public class ProductServiceNotAvailableException extends RuntimeException {
+
+    public ProductServiceNotAvailableException(String message) {
+        super(message);
+    }
+}
