+++ feign/src/main/java/com/baeldung/core/defaulterrorhandling/config/FeignConfig.java
@@ -1,0 +1,13 @@
+package com.baeldung.core.defaulterrorhandling.config;
+
+import org.springframework.context.annotation.Bean;
+
+import feign.Logger;
+
+public class FeignConfig {
+
+    @Bean
+    Logger.Level feignLoggerLevel() {
+        return Logger.Level.FULL;
+    }
+}
