+++ feign/src/main/java/com/baeldung/core/defaulterrorhandling/model/FormData.java
@@ -1,0 +1,12 @@
+package com.baeldung.core.defaulterrorhandling.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class FormData {
+
+    int id;
+    String name;
+}
