+++ feign/src/main/java/com/baeldung/core/defaulterrorhandling/model/Product.java
@@ -1,0 +1,24 @@
+package com.baeldung.core.defaulterrorhandling.model;
+
+public class Product {
+
+    private String id;
+
+    private String productName;
+
+    private double price;
+
+
+    public String getId() {
+        return id;
+    }
+
+    public String getProductName() {
+        return productName;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+}
