+++ feign/src/main/java/com/baeldung/core/fileupload/config/FeignSupportConfig.java
@@ -1,6 +1,5 @@
-package com.baeldung.cloud.openfeign.fileupload.config;
+package com.baeldung.core.fileupload.config;
 
-import org.springframework.beans.factory.ObjectFactory;
 import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
 import org.springframework.cloud.openfeign.support.SpringEncoder;
 import org.springframework.context.annotation.Bean;
@@ -13,12 +12,7 @@ import feign.form.spring.SpringFormEncoder;
 public class FeignSupportConfig {
     @Bean
     public Encoder multipartFormEncoder() {
-        return new SpringFormEncoder(new SpringEncoder(new ObjectFactory<HttpMessageConverters>() {
-            @Override
-            public HttpMessageConverters getObject() {
-                return new HttpMessageConverters(new RestTemplate().getMessageConverters());
-            }
-        }));
+        return new SpringFormEncoder(new SpringEncoder(() -> new HttpMessageConverters(new RestTemplate().getMessageConverters())));
     }
 
     @Bean
