+++ feign/src/main/java/com/baeldung/core/fileupload/service/UploadResource.java
@@ -1,4 +1,4 @@
-package com.baeldung.cloud.openfeign.fileupload.service;
+package com.baeldung.core.fileupload.service;
 
 import org.springframework.web.multipart.MultipartFile;
 
@@ -9,7 +9,7 @@ import feign.Response;
 
 public interface UploadResource {
     
-    @RequestLine("POST /upload-error")
+    @RequestLine("POST /upload-file")
     @Headers("Content-Type: multipart/form-data")
     Response uploadFile(@Param("file") MultipartFile file);
