+++ spring-cloud-modules/spring-cloud-openfeign/README.md
@@ -5,10 +5,4 @@
 - [Provide an OAuth2 Token to a Feign Client](https://www.baeldung.com/spring-cloud-feign-oauth-token)
 - [Propagating Exceptions With OpenFeign and Spring](https://www.baeldung.com/spring-openfeign-propagate-exception)
 - [Feign Client Exception Handling](https://www.baeldung.com/java-feign-client-exception-handling)
-- [File Upload With Open Feign](https://www.baeldung.com/java-feign-file-upload)
-- [Feign Logging Configuration](https://www.baeldung.com/java-feign-logging)
-- [Retrieve Original Message From Feign ErrorDecoder](https://www.baeldung.com/feign-retrieve-original-message)
-- [RequestLine with Feign Client](https://www.baeldung.com/feign-requestline)
-- [Propagating Exceptions With OpenFeign and Spring](https://www.baeldung.com/spring-openfeign-propagate-exception)
-- [Post form-url-encoded Data with Spring Cloud Feign](https://www.baeldung.com/spring-cloud-post-form-url-encoded-data)
 - [Configuring Spring Cloud FeignClient URL](https://www.baeldung.com/spring-cloud-feignclient-url)
