+++ spring-cloud-modules/spring-cloud-openfeign/pom.xml
@@ -61,23 +61,10 @@
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.github.tomakehurst</groupId>
-            <artifactId>wiremock-jre8</artifactId>
-            <version>${wire.mock.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <version>${lombok.version}</version>
-            <scope>provided</scope>
-        </dependency>
     </dependencies>
 
     <properties>
         <spring-cloud.version>2021.0.3</spring-cloud.version>
-        <wire.mock.version>2.33.2</wire.mock.version>
     </properties>
 
 </project>
