--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/config/FeignConfig.java
@@ -1,13 +1,0 @@
-package com.baeldung.cloud.openfeign.config;
-
-import org.springframework.context.annotation.Bean;
-
-import feign.Logger;
-
-public class FeignConfig {
-
-	@Bean
-	Logger.Level feignLoggerLevel() {
-		return Logger.Level.FULL;
-	}
-}
