--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/customizederrorhandling/config/FeignConfig.java
@@ -1,19 +1,0 @@
-package com.baeldung.cloud.openfeign.customizederrorhandling.config;
-
-import org.springframework.context.annotation.Bean;
-
-import feign.Logger;
-import feign.codec.ErrorDecoder;
-
-public class FeignConfig {
-
-    @Bean
-    Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
-    }
-
-    @Bean
-    ErrorDecoder errorDecoder() {
-        return new CustomErrorDecoder();
-    }
-}
