--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/customizederrorhandling/exception/ProductNotFoundException.java
@@ -1,8 +1,0 @@
-package com.baeldung.cloud.openfeign.customizederrorhandling.exception;
-
-public class ProductNotFoundException extends RuntimeException {
-
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
