--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/customizederrorhandling/exception/ProductServiceNotAvailableException.java
@@ -1,8 +1,0 @@
-package com.baeldung.cloud.openfeign.customizederrorhandling.exception;
-
-public class ProductServiceNotAvailableException extends RuntimeException {
-
-    public ProductServiceNotAvailableException(String message) {
-        super(message);
-    }
-}
