--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/defaulterrorhandling/model/FormData.java
@@ -1,12 +1,0 @@
-package com.baeldung.cloud.openfeign.defaulterrorhandling.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class FormData {
-
-    int id;
-    String name;
-}
