--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/defaulterrorhandling/model/Product.java
@@ -1,24 +1,0 @@
-package com.baeldung.cloud.openfeign.defaulterrorhandling.model;
-
-public class Product {
-
-    private String id;
-
-    private String productName;
-
-    private double price;
-
-
-    public String getId() {
-        return id;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-}
