--- spring-cloud-modules/spring-cloud-openfeign/src/main/java/com/baeldung/cloud/openfeign/fileupload/service/UploadClient.java
@@ -1,18 +1,0 @@
-package com.baeldung.cloud.openfeign.fileupload.service;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.multipart.MultipartFile;
-
-import com.baeldung.cloud.openfeign.fileupload.config.FeignSupportConfig;
-
-@FeignClient(name = "file", url = "http://localhost:8081", configuration = FeignSupportConfig.class)
-public interface UploadClient {
-    @PostMapping(value = "/upload-file", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    String fileUpload(@RequestPart(value = "file") MultipartFile file);
-    
-    @PostMapping(value = "/upload-file-error", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    String fileUploadError(@RequestPart(value = "file") MultipartFile file);
-}
